Carry integer widths as u32

+++ src/capability.rs
@@ -7,9 +7,9 @@ use num_traits::One;
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub(crate) enum Capability {
     /// Every value of the unsigned `bits`-bit type maps to a distinct field element: `2^bits <= p`.
-    UnsignedFits(u8),
+    UnsignedFits(u32),
     /// The two's-complement encoding of the signed `bits`-bit type is injective: `2^bits <= p`.
-    SignedFits(u8),
+    SignedFits(u32),
     /// The modulus has at least `bits` bits.
     FieldBitsAtLeast(u32),
 }
@@ -18,7 +18,7 @@ impl Capability {
     pub(crate) fn holds(&self, modulus: &BigUint) -> bool {
         match self {
             Capability::UnsignedFits(bits) | Capability::SignedFits(bits) => {
-                (BigUint::one() << usize::from(*bits)) <= *modulus
+                (BigUint::one() << *bits) <= *modulus
             }
             Capability::FieldBitsAtLeast(bits) => modulus.bits() >= u64::from(*bits),
         }
